test: split test for calc_bond_angle
from BondAngleInteraction
