+++ test/core/CMakeLists.txt
@@ -26,6 +26,7 @@ set(TEST_NAMES
 
     test_bond_length_interaction
     test_bond_angle_interaction
+    test_bond_angle_utility_function
     test_dihedral_angle_interaction
     test_contact_interaction
     test_global_pair_interaction
